+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## XX.XX.X
 * Added 'UnityWebRequest' as the networking handler for WebGL builds.
+* Added "SetID(string newDeviceId)" function for managing device id changes.
 
 ## 23.12.0
 * Added functionality to allow passing custom or overridden metrics.